+++ CHANGES.rst
@@ -14,6 +14,7 @@
 - Fix exit code of ``repozo`` script in case of verification error.
   For details see `#396 <https://github.com/zopefoundation/ZODB/pull/396>`_.
 
+
 5.8.1 (2023-07-18)
 ==================